triangle drawing functions moved to seperate file
