+++ vpr/src/draw/draw.cpp
@@ -37,6 +37,7 @@
 #include "draw_rr.h"
 #include "draw_xtoy.h"
 #include "draw_toggle_functions.h"
+#include "draw_triangle.h"
 #include "read_xml_arch_file.h"
 #include "draw_global.h"
 #include "intra_logic_block.h"
@@ -1898,73 +1899,6 @@ void draw_reset_blk_color(ClusterBlockId blk_id) {
     t_draw_state* draw_state = get_draw_state_vars();
     draw_state->reset_block_color(blk_id);
 }
-
-/**
- * Draws a small triangle, at a position along a line from 'start' to 'end'.
- *
- * 'relative_position' [0., 1] defines the triangles position relative to 'start'.
- *
- * A 'relative_position' of 0. draws the triangle centered at 'start'.
- * A 'relative_position' of 1. draws the triangle centered at 'end'.
- * Fractional values draw the triangle along the line
- */
-void draw_triangle_along_line(ezgl::renderer* g, ezgl::point2d start, ezgl::point2d end, float relative_position, float arrow_size) {
-    VTR_ASSERT(relative_position >= 0. && relative_position <= 1.);
-    float xdelta = end.x - start.x;
-    float ydelta = end.y - start.y;
-
-    float xtri = start.x + xdelta * relative_position;
-    float ytri = start.y + ydelta * relative_position;
-
-    draw_triangle_along_line(g, xtri, ytri, start.x, end.x, start.y, end.y,
-                             arrow_size);
-}
-
-/* Draws a triangle with it's center at loc, and of length & width
- * arrow_size, rotated such that it points in the direction
- * of the directed line segment start -> end.
- */
-void draw_triangle_along_line(ezgl::renderer* g, ezgl::point2d loc, ezgl::point2d start, ezgl::point2d end, float arrow_size) {
-    draw_triangle_along_line(g, loc.x, loc.y, start.x, end.x, start.y, end.y,
-                             arrow_size);
-}
-
-/**
- * Draws a triangle with it's center at (xend, yend), and of length & width
- * arrow_size, rotated such that it points in the direction
- * of the directed line segment (x1, y1) -> (x2, y2).
- *
- * Note that the parameters are in a strange order
- */
-
-void draw_triangle_along_line(ezgl::renderer* g, float xend, float yend, float x1, float x2, float y1, float y2, float arrow_size) {
-    float switch_rad = arrow_size / 2;
-    float xdelta, ydelta;
-    float magnitude;
-    float xunit, yunit;
-    float xbaseline, ybaseline;
-
-    std::vector<ezgl::point2d> poly;
-
-    xdelta = x2 - x1;
-    ydelta = y2 - y1;
-    magnitude = sqrt(xdelta * xdelta + ydelta * ydelta);
-
-    xunit = xdelta / magnitude;
-    yunit = ydelta / magnitude;
-
-    poly.push_back({xend + xunit * switch_rad, yend + yunit * switch_rad});
-    xbaseline = xend - xunit * switch_rad;
-    ybaseline = yend - yunit * switch_rad;
-    poly.push_back(
-        {xbaseline + yunit * switch_rad, ybaseline - xunit * switch_rad});
-    poly.push_back(
-        {xbaseline - yunit * switch_rad, ybaseline + xunit * switch_rad});
-
-    g->fill_poly(poly);
-}
-
-
 void draw_mux_with_size(ezgl::point2d origin, e_side orientation, float height, int size, ezgl::renderer* g) {
     g->set_color(ezgl::YELLOW);
     auto bounds = draw_mux(origin, orientation, height, g);
