+++ vpr/src/draw/draw_rr.cpp
@@ -24,7 +24,7 @@
 #include "draw_rr.h"
 #include "draw_xtoy.h"
 #include "draw_toggle_functions.h"
-
+#include "draw_triangle.h"
 #include "read_xml_arch_file.h"
 #include "draw_global.h"
 #include "intra_logic_block.h"
